various improvements to the product overview page
