+++ src/components/product-overview/ConstituentDetail.js
@@ -4,11 +4,11 @@ import { colors } from '../../styles/theme';
 
 const ConstituentDetail = ({ name, percentage }) => {
   
-  const progressColor = colors.t1Blue;
-  const progressWidth = `${10 + percentage * 3}%`;
+  const progressColor = '#050A30';
+  const progressWidth = `${10 + percentage * 5}%`;
 
   return (
-    <Box position="relative" borderRadius="md" height="40px" width="100%" bg="gray.200" overflow="hidden">
+    <Box position="relative" borderRadius="md" height="48px" width="100%" bg="#EEEEEE" overflow="hidden">
       <Box 
         height="100%" 
         width={progressWidth} 
@@ -17,7 +17,7 @@ const ConstituentDetail = ({ name, percentage }) => {
       />
       <Box position="absolute" top="0" left="0" right="0" bottom="0" display="flex" justifyContent="space-between" alignItems="center" px={4}>
         <Text fontSize="md" fontWeight="bold" color="white">{name}</Text>
-        <Text fontSize="md" color={colors.gray_content} fontWeight="bold">{percentage} %</Text>
+        <Text fontSize="md" color={'#757575'} fontWeight="bold">{percentage} %</Text>
       </Box>
     </Box>
   );
