+++ src/components/product-overview/index.js
@@ -1,23 +1,25 @@
-import { Box, Text } from '@chakra-ui/react';
+import { Box, Text, Stack } from '@chakra-ui/react';
 
 import ChartSection from './ChartSection.js';
-import InfoSection from './InfoSection.js';
+import AssetAllocation from './AssetAllocation.js';
 import QuickFacts from './QuickFacts.js';
 import Characteristics from './Characteristics.js';
 
 const OverviewAndPerformance = () => {
   return (
-    <Box background="#FCFCFC" p="24px" mb="24px" width={{ base: "95%", lg: "70%" }}>
-      <Text fontSize="lg" fontWeight="bold" mb="4">
+    <Box background="white" p="40px" mb="40px" width={{ base: "95%", lg: "70%" }}>
+      <Text fontSize="xxl" fontWeight="bold" mb="4">
         Ethereum Governance Index
       </Text>
-      <Text fontSize="md" color="gray.600" mb="8">
+      <Text fontSize="lg" color="black" mb="80px">
         The Ethereum Governance Index by DCgen is a market-capitalization weighted benchmark tracking the top thirty governance tokens on Ethereum, offering a dynamic measure of this segment's performance.
       </Text>
-      <ChartSection />
-      <QuickFacts />
-      <Characteristics />
-      <InfoSection />
+      <Stack spacing={32}>
+        <ChartSection />
+        <QuickFacts />
+        <Characteristics />
+        <AssetAllocation />
+      </Stack>
     </Box>
   );
 };
